rename assignment-lib => assignment-contacts to be better descriptive
